Move transform_split_values to the vatools namespace
